Rename connectors directory
