+++ ui/sqlconnectors/GenericDetails.tsx
