+++ ui/sqlconnectors/SQLiteDetails.tsx
