+++ ui/sqlconnectors/SnowflakeDetails.tsx
